remove waterfall cruft; spectrum_plot stores 8-bit signed dB values
spectrum_update() only uses bins 1269 - 1803, for some reason
which is not well explained AFAICT.
https://www.vu2ese.com/index.php/2022/07/20/the-sbitx-the-sdr-for-the-homebrewer/
explains why we use MAX_BINS / 2, I guess, but not why we use
fewer than 1024 of the "good" bins. But one possible explanation
(suggested by copilot, posted by KB2ML on discord) is that the center
bin is 1536 (2048 * 3/4), we have 534 useful bins, and at least they are
symmetric about the center. FFT bin spacing (full 2048 @ 48 kHz): 48000
/ 2048 = 23.4375 Hz. web_get_spectrum() uses 48.875 which is 2 *
23.4375, maybe because we are "only plotting one sideband / decimating /
pairing bins. Spectrum span covered ≈ 534 * 46.875 ≈ 25,031.25 Hz
(~25 kHz). So 534 is: (desired_display_span_Hz) / (46.875 Hz per
displayed bin), rounded to an integer." Errm ok something like that.

I also wouldn't be surprised if the reason for 534 bins is that it
fits reasonably well on the 800-pixel-wide touchscreen, alongside the
FT8 log. But I think we should eventually have higher resolution, to
make zooming the waterfall work better; and I still doubt that if
we start with 25khz, and then chop off the ends because they don't
look like useful data, that we can still claim to have a 25khz
display mode. But maybe I'm missing something about what is getting
chopped off. Then, 1024 or 2048 bins might be a waste of space if we
can't fill them. Or, maybe the sample rate is really supposed to be
96khz: then I guess we start with 48k bandwidth, and chop off the
ends to get down to 25k. That seems plausible: it would be
48k * 267 / 512 = 25031.25 total, 93.75 per bin... off by a
factor of 2 again, so I'm still missing something.

Anyway, given that the values are in dB, int8_t provides good enough
range; and we want to read back 8-bit data from the 9p "spectrum" file,
so let's just define it as a block of signed 8-bit values.

Amends ef1796dcf87506bc07cbeb5a30f78ab99caf0ba2

+++ src/sdr.h
@@ -98,7 +98,7 @@ for the transmission. The data required is the same!
 */
 
 extern float fft_bins[];
-extern int spectrum_plot[];
+extern int8_t spectrum_plot[];
 extern struct filter *ssb;
 
 //vfo definitions
